+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   - `#conditional_level`, `#level`, `#nesting_level` , `#set_level`
   - this allows a more unified handling with `Expression::Base` instances
 
+### Fixed
+
+- fixed interpretation of `+` and `?` following interval quantifiers (`{n,n}`)
+  - treat as chained quantifiers like Ruby does instead of possessive/lazy mode
+
 ## [2.3.1] - 2022-04-24 - [Janosch Müller](mailto:janosch84@gmail.com)
 
 ### Fixed